break out SessionTokenProvider class to it's own file to make code cleaner
