+++ aws_runas/__init__.py
@@ -1,11 +1,10 @@
 import os, time
 import logging
-import json, datetime
 import boto3
 import argparse
 import multiprocessing
 
-from botocore.exceptions import ProfileNotFound
+from .session_token_provider import SessionTokenProvider
 
 try:
   # Python 3
@@ -14,65 +13,7 @@ except ImportError:
   # Python 2
   from ConfigParser import ConfigParser
 
-__VERSION__ = '0.2.0-alpha'
-
-class SessionTokenProvider:
-  CACHE_DIR = os.path.expanduser(os.path.join('~', '.aws'))
-  DT_FORMAT = "%Y-%m-%dT%H:%M:%S%z"
-
-  def __init__(self, profile, mfa_serial):
-    self.profile = profile
-    self.mfa_serial = mfa_serial
-    self.cache_file = os.path.join(self.CACHE_DIR, ".aws_session_token_" + profile)
-
-  def _update_cache(self, data):
-    f = open(self.cache_file, 'w')
-
-    try:
-      json.dump(data, f, default=self._fixup_aws_res)
-    finally:
-      if f:
-        f.close()
-
-  def _fixup_aws_res(self, obj):
-    if isinstance(obj, datetime.datetime):
-      return obj.strftime(self.DT_FORMAT)
-    else:
-      return obj
-
-  def _fetch_cache_token(self):
-    expired = True
-    f = open(self.cache_file, 'r')
-
-    try:
-      tok = json.load(f)
-      exp_time = datetime.datetime.strptime(tok['Credentials']['Expiration'], self.DT_FORMAT)
-      expired = time.mktime(exp_time.utctimetuple()) < time.mktime(datetime.datetime.utcnow().utctimetuple())
-    finally:
-      if f:
-        f.close()
-
-    return (tok, expired)
-
-  def _fetch_fresh_token(self):
-    mfa_token = input("Enter MFA Code: ")
-
-    ses = boto3.Session(profile_name=self.profile)
-    sts = ses.client('sts')
-    res = sts.get_session_token(SerialNumber=self.mfa_serial, TokenCode=mfa_token)
-    self._update_cache(res)
-
-    return res
-
-  def get_credentials(self):
-    expired = True
-    if os.path.isfile(self.cache_file):
-      (res, expired) = self._fetch_cache_token()
-
-    if expired:
-      res = self._fetch_fresh_token()
-
-    return res
+__VERSION__ = '0.2.0-alpha1'
 
 def parse_cmdline():
   p = argparse.ArgumentParser(description='Create an environment for interacting with the AWS API using an assumed role')
@@ -178,7 +119,7 @@ def main():
   logging.getLogger('botocore').setLevel(logging.WARNING)
   logging.getLogger('boto3').setLevel(logging.WARNING)
 
-  ses = boto3.Session(profile_name=args.profile) # Does AssumeRole for us
+  ses = boto3.Session(profile_name=args.profile)
 
   if args.list_mfa:
     c = ses.client('iam')
